asena/requirements.txt
affine==2.3.0
amqp==5.1.0
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cligj==0.7.2
cycler==0.11.0
Django==3.2.12
django-celery-beat==2.2.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
et-xmlfile==1.1.0
fonttools==4.31.2
idna==3.3
jdatetime==4.1.0
kiwisolver==1.4.0
kombu==5.2.4
matplotlib==3.5.1
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
rasterio==1.2.10
requests==2.27.1
scipy==1.8.0
six==1.16.0
snuggs==1.4.7
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
gunicorn==20.1.0
